fix(voice-host): the wearable task writer interpolated user content unguarded
src/voice-host.ts (branch-new, commit cb3a25fb) builds a task file with

    `task: ${task.replace(/\n/g, ' ')}`

Newline-folding only. No confineUserContent(), so a model- or device-supplied
task body could carry a forged header line or a ===FENCE=== into the task file
-- the structural-injection vector the guard exists to close.

injection-guard-sweep caught it and is RED on feat/sutando-server while GREEN
on main, so this is a branch regression, not a pre-existing gap. That branch
receives no CI (PRs run against main), which is why it sat unseen.

NOT fixed by adding a third copy of the guard. Two hand-copies already existed
-- task-bridge.ts and the phone skill's conversation-server.ts -- and they had
already drifted once, recorded in conversation-server's own comment: "main's
version dropped the wrapping". So this extracts the shared owner the code
already named: src/task_body_guard.ts, the TS mirror of task_body_guard.py.

  - src/task_body_guard.ts   NEW: the one TS implementation, exported.
                             Constants MOVED verbatim, not retyped -- dropping
                             one of the 41 header keys would be a silent hole.
  - src/task-bridge.ts       drops its private copy, imports the shared one.
  - src/voice-host.ts        imports it and applies it to the task body.

conversation-server.ts keeps its copy for now: it is a skill, so importing from
src/ crosses the skill boundary where this repo has a documented TS
symlink-resolution hazard. Copies go 2 -> 1, and that one is named here rather
than left silent.

The sweep's own couplings to task-bridge.ts as "the TS home" had to move too:
the header-key parity check now reads task_body_guard.ts, and "confineUserContent
defined in task-bridge" is replaced by three stronger checks -- defined+exported
in the owner, and each adapter DELEGATES. 49 -> 52 guards.

Mutation-verified, and the first round found my own new guard vacuous: deleting
the confinement call from voice-host left the suite GREEN, because "delegates"
only asserted the import. Added an explicit call-site assertion. Now:

  voice-host guard removed    -> 51/52 FAILED
  task-bridge import removed  -> 50/51 FAILED
  shared export removed       -> 50/51 FAILED
  restored                    -> 52/52 passed
